Use v2 of the KISSmetrics JS script
Since 2012, the KISSmetrics JS script was updated to v2 (as indicated in
the URL -> scripts.kissmetrics.com/<api_key>.2.js).  This updates the
script code that's injected with km.js.

Other, minor changes:
- Extract KissmetricsAPI into separate file
- Make :api_key a class attr_accessor rather than a method you need to
override
- Update ruby syntax to > 1.9
